Split typechecker state
