+++ src/typechecker/state.rs
@@ -1,0 +1,69 @@
+pub struct TypeChecker {
+    structs: HashMap<String, StructInfo>,
+    enums: HashMap<String, EnumInfo>,
+    functions: HashMap<String, FuncInfo>,
+    methods: HashMap<String, FuncInfo>, // key: "TypeName.method_name"
+    behaviors: HashMap<String, BehaviorInfo>,
+    behavior_extends: HashMap<String, Vec<BehaviorParentRef>>,
+    behavior_extends_spans: HashMap<String, Span>,
+    behavior_impls: HashSet<(String, String)>,
+    behavior_refs_by_key: HashMap<String, BehaviorParentRef>,
+    generic_functions: HashMap<String, GenericFunctionTemplate>,
+    generic_methods: HashMap<String, GenericFunctionTemplate>,
+    specialized_functions: Vec<TypedFunction>,
+    specializations_seen: HashSet<String>,
+    specialized_types: Vec<TypedTypeDef>,
+    specialized_types_seen: HashSet<String>,
+    type_substitutions: Vec<HashMap<String, Type>>,
+    imports: HashMap<String, Vec<String>>, // imported name -> source module path
+    scopes: Vec<Scope>,
+    diagnostics: Vec<Diagnostic>,
+    current_return_type: Option<Type>,
+    current_self_type: Option<Type>,
+    pending_defers: Vec<TypedExpression>,
+    resolver_backed_collection: bool,
+    resolver_behavior_impl_refs: HashMap<String, VecDeque<BehaviorRefMetadata>>,
+    resolver_behavior_required_refs: HashMap<String, VecDeque<BehaviorRefMetadata>>,
+    resolver_missing_behavior_impl_refs: HashSet<String>,
+    resolver_missing_behavior_required_refs: HashSet<String>,
+}
+
+impl Default for TypeChecker {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl TypeChecker {
+    pub fn new() -> Self {
+        Self {
+            structs: HashMap::new(),
+            enums: HashMap::new(),
+            functions: HashMap::new(),
+            methods: HashMap::new(),
+            behaviors: HashMap::new(),
+            behavior_extends: HashMap::new(),
+            behavior_extends_spans: HashMap::new(),
+            behavior_impls: HashSet::new(),
+            behavior_refs_by_key: HashMap::new(),
+            generic_functions: HashMap::new(),
+            generic_methods: HashMap::new(),
+            specialized_functions: Vec::new(),
+            specializations_seen: HashSet::new(),
+            specialized_types: Vec::new(),
+            specialized_types_seen: HashSet::new(),
+            type_substitutions: Vec::new(),
+            imports: HashMap::new(),
+            scopes: vec![Scope::new()], // global scope
+            diagnostics: Vec::new(),
+            current_return_type: None,
+            current_self_type: None,
+            pending_defers: Vec::new(),
+            resolver_backed_collection: false,
+            resolver_behavior_impl_refs: HashMap::new(),
+            resolver_behavior_required_refs: HashMap::new(),
+            resolver_missing_behavior_impl_refs: HashSet::new(),
+            resolver_missing_behavior_required_refs: HashSet::new(),
+        }
+    }
+}
