+++ tests/docs_truth/repo_hygiene/mod.rs
@@ -24,3 +24,4 @@ mod typechecker_program_checking;
 mod typechecker_resolver_validation;
 mod typechecker_runtime;
 mod typechecker_semantic_validation;
+mod typechecker_state;
